pkp/pkp-lib#12337 centralise log viewer code to own service provider
